reuse chimoney helpers from content script in tests! :tada:
Playwright rocks! Minus their Firefox support
