+++ src/content/keyAutoAdd/chimoney.ts
@@ -6,6 +6,7 @@ import {
   type StepRun as Run,
 } from './lib/keyAutoAdd';
 import { isTimedOut, waitForElement, waitForURL } from './lib/helpers';
+import { getAuthToken, getWalletAddressId } from './lib/helpers/chimoney';
 
 // #region: Steps
 
@@ -96,56 +97,6 @@ const addKey: Run<void> = async ({ publicKey }, { output }) => {
   if (data?.status !== 'success') {
     throw new Error(`Failed to upload public key (${await res.text()})`);
   }
-};
-// #endregion
-
-// #region: Helpers
-const getAuthToken = (): string => {
-  const getFirebaseAuthKey = () => {
-    for (let i = 0; i < sessionStorage.length; i++) {
-      const key = sessionStorage.key(i);
-      if (key?.startsWith('firebase:authUser:')) {
-        return key;
-      }
-    }
-  };
-
-  const key = getFirebaseAuthKey();
-  if (!key) {
-    throw new Error('No Firebase auth key found');
-  }
-  const firebaseDataStr = sessionStorage.getItem(key);
-  if (!firebaseDataStr) {
-    throw new Error('No Firebase auth data found');
-  }
-  const firebaseData: {
-    stsTokenManager: {
-      accessToken: string;
-      refreshToken: string;
-      expirationTime: number;
-    };
-  } = JSON.parse(firebaseDataStr);
-  const token = firebaseData?.stsTokenManager?.accessToken;
-  if (!token) {
-    throw new Error('No Firebase auth token found');
-  }
-  const JWT_REGEX =
-    /^([A-Za-z0-9-_=]{2,})\.([A-Za-z0-9-_=]{2,})\.([A-Za-z0-9-_=]{2,})$/;
-  if (!JWT_REGEX.test(token)) {
-    throw new Error('Invalid Firebase auth token');
-  }
-  return token;
-};
-
-const getWalletAddressId = async (): Promise<string> => {
-  // A Firebase request will set this field eventually. We wait max 6s for that.
-  let attemptToFindWalletAddressId = 0;
-  while (++attemptToFindWalletAddressId < 12) {
-    const walletAddressId = sessionStorage.getItem('walletAddressId');
-    if (walletAddressId) return walletAddressId;
-    await sleep(500);
-  }
-  throw new Error('No walletAddressId found in sessionStorage');
 };
 // #endregion
