+++ tests/e2e/helpers/chimoney.ts
@@ -1,4 +1,8 @@
 import type { Page } from '@playwright/test';
+import {
+  getAuthToken,
+  getWalletAddressId,
+} from '@/content/keyAutoAdd/lib/helpers/chimoney';
 
 export const URLS = {
   get login() {
@@ -25,58 +29,8 @@ export async function revokeKey(page: Page, jwkKeyId: string) {
   await page.goto(URLS.keyPage);
 
   // The auth token changes on each page load?!
-  const authToken = await page.evaluate(() => {
-    // Same as from src/content/keyAutoAdd/chimoney.ts
-    const getAuthToken = (): string => {
-      const getFirebaseAuthKey = () => {
-        for (let i = 0; i < sessionStorage.length; i++) {
-          const key = sessionStorage.key(i);
-          if (key?.startsWith('firebase:authUser:')) {
-            return key;
-          }
-        }
-      };
-
-      const key = getFirebaseAuthKey();
-      if (!key) {
-        throw new Error('No Firebase auth key found');
-      }
-      const firebaseDataStr = sessionStorage.getItem(key);
-      if (!firebaseDataStr) {
-        throw new Error('No Firebase auth data found');
-      }
-      const firebaseData: {
-        stsTokenManager: {
-          accessToken: string;
-          refreshToken: string;
-          expirationTime: number;
-        };
-      } = JSON.parse(firebaseDataStr);
-      const token = firebaseData?.stsTokenManager?.accessToken;
-      if (!token) {
-        throw new Error('No Firebase auth token found');
-      }
-      const JWT_REGEX =
-        /^([A-Za-z0-9-_=]{2,})\.([A-Za-z0-9-_=]{2,})\.([A-Za-z0-9-_=]{2,})$/;
-      if (!JWT_REGEX.test(token)) {
-        throw new Error('Invalid Firebase auth token');
-      }
-      return token;
-    };
-    return getAuthToken();
-  });
-
-  const walletAddressId = await page.evaluate(async () => {
-    // Same as from src/content/keyAutoAdd/chimoney.ts
-    // A Firebase request will set this field eventually. We wait max 6s for that.
-    let attemptToFindWalletAddressId = 0;
-    while (++attemptToFindWalletAddressId < 12) {
-      const walletAddressId = sessionStorage.getItem('walletAddressId');
-      if (walletAddressId) return walletAddressId;
-      await new Promise((res) => setTimeout(res, 500));
-    }
-    throw new Error('No walletAddressId found in sessionStorage');
-  });
+  const authToken = await page.evaluate(getAuthToken);
+  const walletAddressId = await page.evaluate(getWalletAddressId);
 
   const keyId = await page.evaluate(
     async ({ walletAddressId, authToken, jwkKeyId }) => {
